Added some more references to the discussion/recommendations
I can’t find all the “so called” papers reviewer 3 alluded to, but in
general I found a couple, but they don’t link that well to these
recommendations.  Perhaps there are such papers but I couldn’t find
them in this short time period, perhaps before the presentation it
would be worth it to search more!